turso-agent/requirements.txt
# Turso Agent - Dependências
# Agente Especialista em Turso Database & MCP Integration

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Turso Database
libsql-client>=0.0.10

# MCP Integration
mcp>=0.0.1

# Async support
aiohttp>=3.9.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# Performance
orjson>=3.9.0
ujson>=5.8.0 